readRect and readPixel fixes
It appears like the reading of data we should have been passing 0x3f
bytes to display instead of 0.
readRect looks like it is now returning valid data.

Also added quick change to readPixel, that if we are on SPI buss whoes
queue size < 4, call off to readRect to handle it.  We need to not push
several things into queue and expect to be able to pop off 3 elements to
calcualte a pixel value...