Fix opaq FB text output end x code
There was a bug in my Frame buffer Opaque text code that did not update
what the actual x was, when outputting bits so when it blanked the reset
of the character to left it blanked to much.

Also updated some of the non-transparent code to not duplicate a bunch
of code depending on if  b=0, but instead simply set N=1 else read in
the N count.