writeRect1NBPP, writeRect2NBPP, writeRect4NBPP - combined
Got tired of lots of very similar code, so decided to write an
implemention that works for 1, 2, 4, 8 bits per pixel, and then so far
for 1, 2, 4 versions they simply call this one.

This one has been updated to hopefully properly handle clipping.  My
test case tries several combinations, but is not exhaustive.