Add Option to enable FRAMEBUFFER  or not...
Put all of the Frame buffer code under #ifdef.  By default the code is
enabled on T3.5 and T3.6 and disabled otherwise.  Put in some negative
test that if a don't use is defined, it won't enable on 3.5/3.6
either... Hopefully will be able to use with Frank's per program
settings