Add ability to autochoose SPIN1, SPIN2
added code that if the SPI pins are not valid for the SPIN object passed
in and SPI1 and SPI2 are valid try seeing if the pins are valid for SPI1
or SPI2 and automatically choose.

Also added error printing when things fail  so you might have a clue.
why things did not work