Fix issues with standalone CS pin not controlled by the SPI module
readPixel() and readRect() were not using beginSPITransaction(),
so the CS pin was not set low correctly during the transaction.
To fix it, beginSPITransaction() needs to allow choosing a clock
speed for reading, which is done with an optional clock parameter.

Also, the CS pin was never set HIGH during begin(), so the display
could incorrectly receive commands intended for other devices.