WIP: Try an UPdateScreen using DMA
Experimenting with adding DMA capability to do the full screen update.
Goal is to be able to start the screen update and then let the code go
and do something else while the update is happening.

WIP - Test program appears to work... Currently different API to use
DMA.  Once working may combine or rename...

Currently hard coded for SPI0, need to factor for SPI1 and maybe SPI2...
But that may not work on T3.5 if I remember correctly...