Add Support for running a Teensy-LC with this library.
Experiment to see if it makes sense to have a higher speed version for
the Teensy LC.  The TLC does not have support for encoding DC into the
queues, but does speed things up to try to keep things double buffered.

Since this library runs using SPIN library it also gives you the
capability of using SPI1 as well.

I believe most everything is implemented in this version.  Obviously
the off screen frame rectangle support is not implemented for this
board as there is not enough memory to support it.