Fix Hang on Async update
Hopefully fix the hang for doing dma updates.

There was an issue that when the DMA completed, I tried to push a NULL command onto the SPI queue, which may not have room (DMA filled it).  So wait until there is room before doing this.

Also added conditional Debug digitalWrite code that was called in the start, the dma end and wait functions.