Adafruit font - Support
Added support to allow you to include Adafruit fonts into your sketch
that uses this library.  There is also a version of the setFont, that
sees that you are using one of these fonts and sets up to use it.

These fonts act differently than our other fonts, in that their drawn
from a logical base point, instead of from the upper left hand corner.

The transparent output code is more or less identical to what is in the
GFX library.  There are some speed improvements that can be made, but so
far they still simply deal with one logial pixel at a time.

I also brought in the later GFX ability to set the text size differently
for both X and Y range.

Unlike Adafruit - I added an ability to draw the characters Opaque.
There are some minor issues with it, in that it has to guess on range of
area to include in the opaque clearing.  Sort of obvious in X direction,
but Y direction figured out max extent for the specific font.

Also with some of these font's, there can be some pixels that overlap.
So far the new character will tromp on the overlap.  May add code to try
to see which bits should be preserved, but so far I have not done so.

The code supports clipping and offsets of these GFX characters as well.