T3.6 - Add DMA Callback capbilities.
Like T4 update.  allows callback on full and half frame updates.

As part of this I reworked part of the DMA code, which I think makes it a bit cleaner.

Before in NON-continuous mode it had the DMASettings chain 0->1->2 and kills off 3.  It starts off outputting the 2nd pixel in 0 as it needed us to do a PUSH to properly set the proper 32 bits of the PUSHR register...
With Continuous mode, I had 0->1->2->3->0...  where 3 only output 1 pixel...
Now 3 is setup to be sames as 0 except it starts at first pixel and for the full size, so loop is: 0->1->2->3->1...
Just feels cleaner.  So I don't have to muck with chain.  I always setup for interrupt and completion on 2.  Each time I start I just make sure to copy 0->TX...

So tried with my DMA clip test sketch plus with the continuous test sketch...